Added Notes, Note Holder, and Table Numbers
I added all the assets blank, as well as already numbered, because I'm not sure which one you guys will need. Feel free to also not use the numbers I made, and use a pixel font found online instead. Whatever is easiest for you! The idle for the note holder is 120ms and the swing is 100ms. They can also both be 120 if that makes it easier.